update: 取得した予定を表示
