+++ src/components/Clock.tsx
@@ -23,30 +23,6 @@ const Clock: React.FC<Props> = ({date}) => {
           <span>{getMinutes(date)}:</span>
           <span>{getSeconds(date)}</span>
         </div>
-        <div className="w-full rounded overflow-hidden shadow-lg mx-6 my-4">
-          <div className="px-6 py-4">
-            <div className="font-bold text-xl mb-2">
-              ペアプロ<time>{new Date().getTime()}</time>
-            </div>
-            <p className="text-gray-700 text-base">electron！</p>
-          </div>
-        </div>
-        <div className="w-full rounded overflow-hidden shadow-lg mx-6 my-4">
-          <div className="px-6 py-4">
-            <div className="font-bold text-xl mb-2">
-              ペアプロ<time>{new Date().getTime()}</time>
-            </div>
-            <p className="text-gray-700 text-base">electron！</p>
-          </div>
-        </div>
-        <div className="w-full rounded overflow-hidden shadow-lg mx-6 my-4">
-          <div className="px-6 py-4">
-            <div className="font-bold text-xl mb-2">
-              ペアプロ<time>{new Date().getTime()}</time>
-            </div>
-            <p className="text-gray-700 text-base">electron！</p>
-          </div>
-        </div>
       </div>
     </div>
   );
