+++ src/components/Task.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import {Props} from '../types/Task.types';
+
+const Task: React.FC<Props> = ({summary, start}) => {
+  return (
+    <div className="w-full rounded overflow-hidden shadow-lg mx-6 my-4">
+      <div className="px-6 py-4">
+        <div className="font-bold text-xl mb-2">
+          {summary}
+          <p>{start}</p>
+        </div>
+        <p className="text-gray-700 text-base"></p>
+      </div>
+    </div>
+  );
+};
+
+export default Task;
