--- src/types/Calendar.types.ts
@@ -1,4 +1,0 @@
-export type Props = {
-  dateFrom: string;
-  dateTo: string;
-};
