+++ src/types/ShowTask.types.ts
@@ -1,0 +1,4 @@
+export type Events = Array<{
+  summary: string;
+  start: {dateTime: string};
+}>;
