+++ src/types/Task.types.ts
@@ -1,0 +1,4 @@
+export type Props = {
+  summary: string;
+  start: string;
+};
